+++ CHANGELOG.md
@@ -9,8 +9,8 @@ Version 0.27
 - improved `CREATE TYPE` for base types, shell types support
 - improved dropping of indexes which are really constraints
 - compatibility improvements with old postgres versions
+- `comments` option to show all comments, even null ones, these are now omitted by default if null
 STILL TO DO:
-- `comments` option to show all comments, even NULL ones
 - `noconstraints` + `lite` options should omit constraints
 - `owner` and `noowner` and `nodcl` for serial sequences
 - check if `script` option works ok for sequences